divide actions and reducers
