+++ src/components/SanitaryRegister/SanitaryRegisterForm.js
@@ -49,6 +49,7 @@ function SanitaryRegisterForm({
               type="checkbox"
               name="question1SymptomsCovid"
               value={question1SymptomsCovid}
+              checked={question1SymptomsCovid}
               onChange={handleChange}
             >
             </Checkbox>
@@ -61,6 +62,7 @@ function SanitaryRegisterForm({
               type="checkbox"
               name="question2ContactWithPeople"
               value={question2ContactWithPeople}
+              checked={question2ContactWithPeople}
               onChange={handleChange}
             >
             </Checkbox>
@@ -73,6 +75,7 @@ function SanitaryRegisterForm({
               type="checkbox"
               name="question3InternationalTravel"
               value={question3InternationalTravel}
+              checked={question3InternationalTravel}
               onChange={handleChange}
             >
             </Checkbox>
@@ -85,6 +88,7 @@ function SanitaryRegisterForm({
               type="checkbox"
               name="question4HealthWorker"
               value={question4HealthWorker}
+              checked={question4HealthWorker}
               onChange={handleChange}
             >
             </Checkbox>
