+++ src/store/actions/SanitaryRegister.actions.js
@@ -1,18 +1,17 @@
 import axios from 'axios'
-
-const CREATE_SANITARY_REGISTER = 'CREATE_SANITARY_REGISTER';
-const FAILURED_SANITARY_REGISTER = 'FAILURED_SANITARY_REGISTER';
-const LOADING = 'LOADING';
-const FINISHED_LOADING = 'FINISHED_LOADING';
-const GET_DATA_FORM = 'GET_DATA_FORM';
-
-const QUESTION1 = 'QUESTION1';
-const QUESTION2 = 'QUESTION2';
-const QUESTION3 = 'QUESTION3';
-const QUESTION4 = 'QUESTION4';
-const TEMPERATURE = 'TEMPERATURE';
-
-const CANCEL_QUESTION1 = 'CANCEL_QUESTION1';
+import { 
+  CREATE_SANITARY_REGISTER,
+  FAILURED_SANITARY_REGISTER,
+  LOADING,
+  FINISHED_LOADING,
+  GET_DATA_FORM,
+  QUESTION1,
+  QUESTION2,
+  QUESTION3,
+  QUESTION4,
+  TEMPERATURE,
+  CANCEL_QUESTION1,
+} from '../reducers/SanitaryRegister.reducers'
 
 export function getQuestionOne( payload ) {
   return function( dispatch ) {
@@ -47,7 +46,7 @@ export function getTemperature( payload ) {
 export function createSanitaryRegister(data) {
   const { question1SymptomsCovid, question2ContactWithPeople, question3InternationalTravel, question4HealthWorker, temperature } = data
   return async function (dispatch) {
-    dispatch({ type: LOADING })
+    // dispatch({ type: LOADING })
     try {
       const token = localStorage.getItem('token');
       const { data } = await axios({
@@ -113,7 +112,6 @@ export function getData() {
       dispatch({ type: QUESTION3, payload: data.question3InternationalTravel })
       dispatch({ type: QUESTION4, payload: data.question4HealthWorker })
       dispatch({ type: TEMPERATURE, payload: data.temperature })
-      console.log(data)
     })
     .catch(err => {
       dispatch({ 
@@ -121,81 +119,5 @@ export function getData() {
         payload: 'Lo sentimos, vuelve a recargar la página para cargar la información'
       })
     })
-  }
-};
-
-
-export const initialState = {
-  question1SymptomsCovid: false,
-  question2ContactWithPeople: false,
-  question3InternationalTravel: false,
-  question4HealthWorker: false,
-  temperature: '',
-  errorSubmittion: '',
-  message: '',
-  messageTemperature: '',
-  loading: false,
-  id: '',
-};
-
-export function sanitaryRegisterReducer( state = initialState, action) {
-  switch (action.type) {
-    case LOADING: 
-      return {
-        ...state,
-        loading: true,
-      }
-    case FINISHED_LOADING:
-      return {
-        ...state,
-        loading: false,
-      }
-    case CREATE_SANITARY_REGISTER:
-      return {
-        ...state,
-        message: action.payload,
-      };
-    case FAILURED_SANITARY_REGISTER:
-      return {
-        ...state,
-        errorSubmittion: action.payload,
-      }
-    case QUESTION1:
-      return {
-        ...state,
-        question1SymptomsCovid: action.payload,
-      };
-    case QUESTION2:
-      return {
-        ...state,
-        question2ContactWithPeople: action.payload,
-      };
-    case QUESTION3:
-      return {
-        ...state,
-        question3InternationalTravel: action.payload,
-      };
-    case QUESTION4:
-      return {
-        ...state,
-        question4HealthWorker: action.payload,
-      };
-    case TEMPERATURE:
-      return {
-        ...state,
-        temperature: action.payload,
-      };
-    case CANCEL_QUESTION1:
-      return {
-        ...state,
-        question1SymptomsCovid: action.payload,
-      }
-    case GET_DATA_FORM:
-      return {
-        ...state,
-        id: action.payload
-      }
-    default:
-      return state;
   }
 };
